chore: composable renamed, new packages added into composable
